fix(postcss): extract plugins configurations
